Add functionality to edit and delete movies
Added a new Javascript file 'editAndDeleteMovies.js' which fetches movie data from a server and displays it in a table with 'Edit' and 'Delete' buttons. Also, added a new HTML page 'editarpelicula.html' for movie editing. When clicked, the 'Edit' button on particular movie row will redirect to the new page with the movie ID in the URL, and 'Delete' button will remove the movie after a confirmation.