Add movie list page and associated editing script
A new page was created named `listadodepeliculas.html` for listing movies. Functionality for movie editing was also added in the script `movieEditFormHandler.js`. The script fetches movie data using a GET request, populates the form for editing, and also processes the form submission with a PATCH request for updating the selected movie.